Removed developer.default.yaml file
For our simple (open source) projects, the settings file is only
necessary to save the build artifacts to another location. If you are
building locally, then you can run rake without a settings file at all
and it will pick up the default settings which are the same as in the
developer.default.yaml file. To run without a settings file:

    rake default

or just:

    rake